(BSR) feat(cookies): move PrivacyPolicy from FirstLogin to Cookies folder
